# src/main/resources/static/dape-api-contract.yml
openapi: 3.0.3
info:
  title: DAPE
  description: DAPE API contract - Declaração de Apostas Esportivas
  contact:
    email: efchaves@uolinc.com
  version: 1.0.1
servers:
  - url: http://localhost:8080/dape
    description: Servidor local
...
paths:
  /bet:
    post:
      tags:
        - bet
      summary: Cadastrar aposta
      description: Realizar o cadastro de uma nova aposta na base de dados
      operationId: cadastrarAposta
      requestBody:
        description: Cadastro de uma nova aposta no banco de dados
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BetCreateRequest'
        required: true
      responses:
        '201':
          description: Created, Aposta cadastrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetCreateResponse'
        '400':
          description: Bad Request, Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - bet
      summary: Consultar apostas
      description: Consultar as apostas disponíveis no banco de dados
      operationId: consultarAposta
      parameters:
        - name: page
          in: query
          required: false
          explode: true
          schema:
            type: integer
            format: int32
            example: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            example: 5
        - name: bet_status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StatusEnum'
        - name: dat_created
          in: query
          required: false
          schema:
            type: string
            format: date-time
            example: "2025-07-21T15:16:10.877543705"
        - name: dat_updated
          in: query
          required: false
          schema:
            type: string
            format: date-time
            example: "2025-07-21T15:16:10.877543705"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetListResponse'
        '400':
          description: Bad Request, Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    BetCreateRequest:
      required:
        - num_odd
        - des_bet
      type: object
      properties:
        num_odd:
          type: number
          format: double
          example: 2.19
        des_bet:
          type: string
          example: "Vitória do Botafogo"
      xml:
        name: BetCreateRequest
    BetCreateResponse:
      type: object
      properties:
        idt_bet:
          type: integer
          format: int32
          example: 2
        des_bet:
          type: string
          example: "Vitória do Botafogo"
        num_odd:
          type: number
          format: double
          example: 2.19
        dat_created:
          type: string
          format: date-time
          example: "2025-07-21T15:16:10.877543705"
        dat_updated:
          type: string
          format: date-time
          example: "2025-07-21T15:16:10.877543705"
        bet_status:
          $ref: '#/components/schemas/StatusEnum'
        flg_selected:
          type: integer
          format: int32
          example: 0
      xml:
        name: BetCreateRequest
...
    BetListResponse:
      type: object
      properties:
        size:
          type: integer
          format: int32
          example: 10
        total_size:
          type: integer
          format: int32
          example: 1000
        page:
          type: integer
          format: int32
          example: 0
          default: 0
        total_pages:
          type: integer
          format: int32
          example: 100
        content:
          type: array
          items:
            type: object
            properties:
              num_odd:
                type: number
                format: double
                example: 2.19
              des_bet:
                type: string
                example: "Vitória do Botafogo"
              bet_status:
                $ref: '#/components/schemas/StatusEnum'
              dat_created:
                type: string
                format: date-time
                example: "2025-07-21T15:16:10.877543705"
              dat_updated:
                type: string
                format: date-time
                example: "2025-07-21T15:16:10.877543705"
      xml:
        name: BetListResponse
...
    StatusEnum:
      type: string
      enum:
        - PENDING
        - GREEN
        - RED
      xml:
        name: StatusEnum
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Mensagem de erro"
      xml:
        name: ErrorResponse